+++ django_google_maps_widget/templates/django_google_maps_widget/widgets/map_widget.html
