+++ django_google_maps_widget/tests/test_app/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from django_google_maps.fields import AddressField, GeoLocationField
+from django_google_maps_widget.fields import AddressField, GeoLocationField
 
 
 class Person(models.Model):
