+++ sample/admin.py
@@ -2,8 +2,8 @@ from django.conf import settings
 from django.contrib import admin
 from django.forms.widgets import TextInput
 
-from django_google_maps.widgets import GoogleMapsAddressWidget
-from django_google_maps.fields import AddressField, GeoLocationField
+from django_google_maps_widget.widgets import GoogleMapsAddressWidget
+from django_google_maps_widget.fields import AddressField, GeoLocationField
 
 from sample import models
