+++ sample/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 from django.conf import settings
 from sample.models import SampleModel
-from django_google_maps.widgets import GoogleMapsAddressWidget
+from django_google_maps_widget.widgets import GoogleMapsAddressWidget
 
 
 class SampleForm(forms.ModelForm):
